Move samples, sample_type, and seed props to base class
