+++ csdms/dakota/tests/test_method_base_uq.py
@@ -1,6 +1,6 @@
 """Tests for csdms.dakota.method.base.UncertaintyQuantificationBase."""
 
-from nose.tools import raises, assert_true, assert_equal
+from nose.tools import raises, assert_true, assert_equal, assert_is_none
 from csdms.dakota.method.base import UncertaintyQuantificationBase
 
 
@@ -49,6 +49,72 @@ def test_basis_polynomial_family_fails_if_unknown_type():
     c.basis_polynomial_family = value
 
 
+def test_get_samples():
+    """Test getting the samples property."""
+    assert_true(type(c.samples) is int)
+
+
+def test_set_samples():
+    """Test setting the samples property."""
+    m = Concrete()
+    samples = 42
+    m.samples = samples
+    assert_equal(m.samples, samples)
+
+
+@raises(TypeError)
+def test_set_samples_fails_if_float():
+    """Test that the samples property fails with a float."""
+    m = Concrete()
+    samples = 42.0
+    m.samples = samples
+
+
+def test_get_sample_type():
+    """Test getting the sample_type property."""
+    assert_true(c.sample_type == 'random' or c.sample_type == 'lhs')
+
+
+def test_set_sample_type():
+    """Test setting the sample_type property."""
+    m = Concrete()
+    sample_type = 'lhs'
+    m.sample_type = sample_type
+    assert_equal(m.sample_type, sample_type)
+
+
+@raises(TypeError)
+def test_set_sample_type_fails_if_not_lhs_or_random():
+    """Test that the sample_type property fails with unknown type."""
+    m = Concrete()
+    sample_type = 'mcmc'
+    m.sample_type = sample_type
+
+
+def test_get_seed():
+    """Test getting the seed property."""
+    if c.seed is not None:
+        assert_true(type(c.seed) is int)
+    else:
+        assert_is_none(c.seed)
+
+
+def test_set_seed():
+    """Test setting the seed property."""
+    m = Concrete()
+    seed = 42
+    m.seed = seed
+    assert_equal(m.seed, seed)
+
+
+@raises(TypeError)
+def test_set_seed_fails_if_float():
+    """Test that the seed property fails with a float."""
+    m = Concrete()
+    seed = 42.0
+    m.seed = seed
+
+
 def test_str_special():
     """Test type of __str__ method results."""
     s = str(c)
@@ -59,4 +125,4 @@ def test_str_length():
     """Test the default length of __str__."""
     s = str(c)
     n_lines = len(s.splitlines())
-    assert_equal(n_lines, 3)
+    assert_equal(n_lines, 5)
