+++ csdms/dakota/tests/test_method_polynomial_chaos.py
@@ -153,72 +153,6 @@ def test_set_response_levels_fails_if_scalar():
     m.response_levels = pt
 
 
-def test_get_samples():
-    """Test getting the samples property."""
-    assert_true(type(x.samples) is int)
-
-
-def test_set_samples():
-    """Test setting the samples property."""
-    m = PolynomialChaos()
-    samples = 42
-    m.samples = samples
-    assert_equal(m.samples, samples)
-
-
-@raises(TypeError)
-def test_set_samples_fails_if_float():
-    """Test that the samples property fails with a float."""
-    m = PolynomialChaos()
-    samples = 42.0
-    m.samples = samples
-
-
-def test_get_sample_type():
-    """Test getting the sample_type property."""
-    assert_true(x.sample_type == 'random' or x.sample_type == 'lhs')
-
-
-def test_set_sample_type():
-    """Test setting the sample_type property."""
-    m = PolynomialChaos()
-    sample_type = 'lhs'
-    m.sample_type = sample_type
-    assert_equal(m.sample_type, sample_type)
-
-
-@raises(TypeError)
-def test_set_sample_type_fails_if_not_lhs_or_random():
-    """Test that the sample_type property fails with unknown type."""
-    m = PolynomialChaos()
-    sample_type = 'mcmc'
-    m.sample_type = sample_type
-
-
-def test_get_seed():
-    """Test getting the seed property."""
-    if x.seed is not None:
-        assert_true(type(x.seed) is int)
-    else:
-        assert_is_none(x.seed)
-
-
-def test_set_seed():
-    """Test setting the seed property."""
-    m = PolynomialChaos()
-    seed = 42
-    m.seed = seed
-    assert_equal(m.seed, seed)
-
-
-@raises(TypeError)
-def test_set_seed_fails_if_float():
-    """Test that the seed property fails with a float."""
-    m = PolynomialChaos()
-    seed = 42.0
-    m.seed = seed
-
-
 def test_str_special():
     """Test type of __str__ method results."""
     s = str(x)
